requirements.txt
flask
tensorflow
numpy
Pillow
